requirements.txt
django==4.2
django-environ==0.10.0
djangorestframework==3.14.0